📦️ Add GUI
Finally! The first version of a working GUI, The "Write Out To File"
option after password generating still doesn't work, but who cares? 🙂

🖊️ Edit functions' descriptions

Removed args from description, because it was not necessary...

📥️ Implement type pre-check for gen* functions

Functions are now going to return error string when it cannot convert
input to int type.

🖼️ Cool looking icon.

That's it...?